# abacusflow-portal/abacusflow-portal-web/src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: abacusflow api
  version: 0.0.1
  description: 小算盘服务接口文档
servers:
  - url: http://localhost:8080
...
paths:
  /users:
    get:
      operationId: listBasicUsers
      tags:
        - user
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BasicUser"
    post:
      operationId: addUser
      tags:
        - user
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserInput"
      responses:
        "200":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags:
        - user
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags:
        - user
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInput"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags:
        - user
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: User deleted
...
components:
  schemas:
...
    User:
      type: object
      required:
        - id
        - name
        - age
        - nick
        - roleIds
        - enabled
        - locked
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: 用户唯一标识
        name:
          type: string
          description: 用户真实姓名
        sex:
          $ref: "#/components/schemas/Sex"
        age:
          type: integer
          description: 年龄
        nick:
          type: string
          description: 昵称
        roleIds:
          type: array
          items:
            type: integer
            format: int64
          description: 用户角色列表
        enabled:
          type: boolean
          description: 账户是否启用
        locked:
          type: boolean
          description: 账户是否锁定
        createdAt:
          type: integer
          format: int64
          description: 创建时间戳
        updatedAt:
          type: integer
          format: int64
          description: 更新时间戳
    BasicUser:
      type: object
      required:
        - id
        - name
        - nick
        - enabled
        - locked
      properties:
        id:
          type: integer
          format: int64
          description: 用户ID
        name:
          type: string
          description: 用户姓名
        nick:
          type: string
          description: 昵称
        enabled:
          type: boolean
          description: 是否启用
        locked:
          type: boolean
          description: 是否锁定
    CreateUserInput:
      type: object
      required:
        - name
        - nick
      properties:
        name:
          type: string
          description: 账户名
        sex:
          $ref: "#/components/schemas/Sex"
        age:
          type: integer
          minimum: 1
          maximum: 200
          description: 年龄
        nick:
          type: string
          description: 昵称
    UpdateUserInput:
      type: object
      properties:
        sex:
          $ref: "#/components/schemas/Sex"
        age:
          type: integer
          minimum: 1
          maximum: 200
          description: 年龄
        nick:
          type: string
          description: 昵称

    # ============= Supplier schemas =============
...
    Sex:
      type: string
      description: 性别
      enum:
        - male
        - female
      x-enum-varnames:
        - male
        - female
      x-enum-descriptions:
        - 男
        - 女
...